Remove extraneous tests
These tests covered an edge case that was not specified as part of the project requirements. I broke them while working on another feature that is part of the project requirements. The deadline is coming up fast... so I decided to remove the extraneous tests instead of spending limited time to fix them. Moving forward.